Rearrange properties to be compatible with modules, reducing warnings
These are warnings that aren't enabled by default, but which can
sometimes pick up real issues.

+++ SpiNNaker-allocserv/src/main/java/uk/ac/manchester/spinnaker/alloc/bmp/FirmwareLoader.java
@@ -381,7 +381,7 @@ public class FirmwareLoader {
 
 	private void updateFlashData(ByteBuffer data)
 			throws ProcessException, IOException {
-		data.putInt(CRC_OFFSET, (int) ~crc(data, 0, CRC_OFFSET));
+		data.putInt(CRC_OFFSET, ~crc(data, 0, CRC_OFFSET));
 		data.position(0);
 		MemoryLocation fb = txrx.getSerialFlashBuffer(board);
 		txrx.writeBMPMemory(board, fb, data);
